writeback: reduce per-bdi dirty threshold ramp up time
Reduce the dampening for the control system, yielding faster
convergence.

Currently it converges at a snail's pace for slow devices (in order of
minutes).  For really fast storage, the convergence speed should be
fine.

It makes sense to make it reasonably fast for typical desktops.

After patch, it converges in ~10 seconds for 60MB/s writes and 3GB mem.
So expect ~1s for a fast 600MB/s storage on 3GB mem, or ~4s on 12GB mem,
which seems reasonable.

$ while true; do grep BdiDirtyThresh /debug/bdi/btrfs-1/stats; sleep 1; done
0  BdiDirtyThresh:              0 kB
1  BdiDirtyThresh:         198272 kB
2  BdiDirtyThresh:         390500 kB
3  BdiDirtyThresh:         440188 kB
4  BdiDirtyThresh:         480140 kB
5  BdiDirtyThresh:         524064 kB
6  BdiDirtyThresh:         538848 kB
7  BdiDirtyThresh:         553020 kB
8  BdiDirtyThresh:         555848 kB
9  BdiDirtyThresh:         559956 kB
10 BdiDirtyThresh:         562248 kB
11 BdiDirtyThresh:         562880 kB
12 BdiDirtyThresh:         564104 kB
13 BdiDirtyThresh:         564156 kB
14 BdiDirtyThresh:         564256 kB
15 BdiDirtyThresh:         564256 kB

CC: Peter Zijlstra <a.p.zijlstra@chello.nl>
CC: Richard Kennedy <richard@rsk.demon.co.uk>
Signed-off-by: Wu Fengguang <fengguang.wu@intel.com>