mm: compaction: Use the LRU to get a hint on where compaction should start
The end of the LRU stores the oldest known page. Compaction on the other
hand always starts scanning from the start of the zone. This patch uses
the LRU to hint to compaction where it should start scanning from. This
means that compaction will at least start with some old pages reducing
the impact on running processes and reducing the amount of scanning. The
check it makes is racy as the LRU lock is not taken but it should be
harmless as we are not manipulating the lists without the lock.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>