writeback: start background writeback earlier
It's possible for some one to suddenly eat lots of memory,
leading to sudden drop of global dirty limit. So a dirtier
task may get hard throttled immediately without some previous
balance_dirty_pages() call to invoke background writeback.

In this case we need to check for background writeback earlier in the
loop to avoid stucking the application for very long time. This was not
a problem before the IO-less balance_dirty_pages() because it will try
to write something and then break out of the loop regardless of the
global limit.

Another scheme this check will help is, the dirty limit is too close to
the background threshold, so that someone manages to jump directly
beyond the pause threshold (background+dirty)/2.

Simplify the laptop mode handling by the way, since we've kick
background writeback inside the balance_dirty_pages() loop on
dirty_exceeded.

Signed-off-by: Wu Fengguang <fengguang.wu@intel.com>