Add Hybrid Adaptive Voltage Scaling (HAVS).
The purpose of HAVS is to minimize the power used by the CPU by determining
and setting the optimal voltage. At the same time, the maximum and minimum voltage by
which HAVS can scale to is fixed to a specified voltage depending on the
CPU frequency in order to prevent scaling to a higher/lower voltage than what is
known as safe at a specified voltage. The optimal voltage is actively determined
for each frequency and temperature. HAVS actively adjusts the CPU voltage as the
CPU frequency and temperature changes.

AVS Author: Dave Estes

HAVS Modification Author: Raymond Golo

Taken From - intersectR's HAVS Kernel Tree