mmc: core: Fix mmc_delayed_work wakelock bug in mmc_rescan()
The mmc_rescan() should extend mmc_delayed_work wakelock only when card
insertion or removal event is detected. But the current code has a bug
because of which it always extends this wakelock for every mmc_rescan()
after card removal. This will be done until the card is detected again.
This patch fixes this bug by extending the wakelock only when the card
is removed and not for every mmc_rescan after card removal.

Change-Id: Ifa44f195abf7aaed777c2de267bd02f09b5fd989
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>